feat: refactor out health-check and provider
