+++ controllers/account.go
@@ -59,13 +59,3 @@ func (c *ApiController) GetAccount() {
 
 	c.ResponseOk(claims, hostname)
 }
-
-func (c *ApiController) GetProviders() {
-	providers, err := casdoorsdk.GetProviders()
-	if err != nil {
-		c.ResponseError(err.Error())
-		return
-	}
-
-	c.ResponseOk(providers)
-}
