+++ web/src/backend/ProviderBackend.js
@@ -14,23 +14,9 @@
 
 import * as Setting from "../Setting";
 
-export function getAccount() {
-  return fetch(`${Setting.ServerUrl}/api/get-account`, {
+export function getProviders() {
+  return fetch(`${Setting.ServerUrl}/api/get-providers`, {
     method: "GET",
-    credentials: "include",
-  }).then(res => res.json());
-}
-
-export function signin(code, state) {
-  return fetch(`${Setting.ServerUrl}/api/signin?code=${code}&state=${state}`, {
-    method: "POST",
-    credentials: "include",
-  }).then(res => res.json());
-}
-
-export function signout() {
-  return fetch(`${Setting.ServerUrl}/api/signout`, {
-    method: "POST",
     credentials: "include",
   }).then(res => res.json());
 }
